requirements.txt
Flask-RESTful
mongoengine
requests
requests-cache
celery-with-mongodb
